# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Meeting_Chat API",
        "description": "Application for free communication",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Application for free communication"
        }
    ],
    "paths": {
        "/interlocutor/{interlocutorCode}": {
            "get": {
                "tags": [
                    "Interlocutors"
                ],
                "summary": "Show interlocutor",
                "description": "Show interlocutor",
                "operationId": "interlocutor",
                "parameters": [
                    {
                        "name": "interlocutorCode",
                        "in": "path",
                        "description": "Interlocutor code",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GetInterlocutor"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/textChat/{roomName}": {
            "get": {
                "tags": [
                    "TextChat"
                ],
                "summary": "Displays interlocutors",
                "description": "Displays interlocutors - with oneOf examples",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "A list of things.",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "roomName",
                        "in": "path",
                        "description": "Room name",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Response_index_messages"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/textChat/message/{roomName}/{interlocutorCode}": {
            "post": {
                "tags": [
                    "TextChat"
                ],
                "summary": "Create message",
                "description": "Create message",
                "parameters": [
                    {
                        "name": "roomName",
                        "in": "path",
                        "description": "Room name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "interlocutorCode",
                        "in": "path",
                        "description": "Interlocutor code",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "content": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "content": "Test text"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GetMessage"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/textChat/{id}": {
            "delete": {
                "tags": [
                    "TextChat"
                ],
                "summary": "Destroy interlocutor",
                "description": "Destroy interlocutor",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Message id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "int": {
                                "summary": "An int value.",
                                "value": "1"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GetMessage"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "GetInterlocutor": {
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "interlocutorName": {
                        "type": "string"
                    },
                    "code": {
                        "type": "string"
                    },
                    "user": {
                        "oneOf": [
                            {
                                "type": "null"
                            },
                            {
                                "$ref": "#/components/schemas/GetUser"
                            }
                        ]
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    }
                },
                "type": "object",
                "example": {
                    "id": 1,
                    "interlocutorName": "Test",
                    "code": "23984y-uioe897-829347g-34879gb23",
                    "user": "null",
                    "created_at": "2023-11-06T15:34:03.000000Z",
                    "updated_at": "2023-11-06T15:34:03.000000Z"
                }
            },
...
            "GetMessage": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "content": {
                        "type": "string"
                    },
                    "interlocutor": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/GetInterlocutor"
                            }
                        ]
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    }
                },
                "type": "object",
                "example": {
                    "id": 1,
                    "content": "Test text",
                    "interlocutor": {
                        "id": 1,
                        "firstName": "Test",
                        "lastName": "User",
                        "gender": "Female",
                        "phone": "+380501234567",
                        "currentAddress": "Beech Creek, PA, Pennsylvania",
                        "permanantAddress": "Arlington Heights, IL, Illinois",
                        "email": "12345@gmail.com",
                        "birthday": "1975-05-21",
                        "email_verified_at": "2023-11-06T15:34:03.000000Z",
                        "avatarPath": "avatar/127486g-2397y4-28397-2387h4.png",
                        "email_verified": false,
                        "created_at": "2023-11-06T15:34:03.000000Z",
                        "updated_at": "2023-11-06T15:34:03.000000Z"
                    },
                    "created_at": "2023-11-06T15:34:03.000000Z",
                    "updated_at": "2023-11-06T15:34:03.000000Z"
                }
            },
            "Response_index_messages": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer"
                                },
                                "content": {
                                    "type": "string"
                                },
                                "interlocutor": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/GetInterlocutor"
                                        }
                                    ]
                                },
                                "created_at": {
                                    "type": "string"
                                },
                                "updated_at": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object",
                "example": {
                    "data": [
                        {
                            "id": 1,
                            "content": "Test text",
                            "interlocutor": {
                                "id": 1,
                                "firstName": "Test",
                                "lastName": "User",
                                "gender": "Female",
                                "phone": "+380501234567",
                                "currentAddress": "Beech Creek, PA, Pennsylvania",
                                "permanantAddress": "Arlington Heights, IL, Illinois",
                                "email": "12345@gmail.com",
                                "birthday": "1975-05-21",
                                "email_verified_at": "2023-11-06T15:34:03.000000Z",
                                "avatarPath": "avatar/127486g-2397y4-28397-2387h4.png",
                                "email_verified": false,
                                "created_at": "2023-11-06T15:34:03.000000Z",
                                "updated_at": "2023-11-06T15:34:03.000000Z"
                            },
                            "created_at": "2023-11-06T15:34:03.000000Z",
                            "updated_at": "2023-11-06T15:34:03.000000Z"
                        },
                        {
                            "id": 2,
                            "content": "Test text",
                            "interlocutor": {
                                "id": 1,
                                "firstName": "Test",
                                "lastName": "User",
                                "gender": "Female",
                                "phone": "+380501234567",
                                "currentAddress": "Beech Creek, PA, Pennsylvania",
                                "permanantAddress": "Arlington Heights, IL, Illinois",
                                "email": "12345@gmail.com",
                                "birthday": "1975-05-21",
                                "email_verified_at": "2023-11-06T15:34:03.000000Z",
                                "avatarPath": "avatar/127486g-2397y4-28397-2387h4.png",
                                "email_verified": false,
                                "created_at": "2023-11-06T15:34:03.000000Z",
                                "updated_at": "2023-11-06T15:34:03.000000Z"
                            },
                            "created_at": "2023-11-06T15:34:03.000000Z",
                            "updated_at": "2023-11-06T15:34:03.000000Z"
                        }
                    ]
                }
            },
...
            "GetUser": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "gender": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "currentAddress": {
                        "type": "string"
                    },
                    "permanantAddress": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "birthdey": {
                        "type": "string"
                    },
                    "email_verified": {
                        "type": "string"
                    },
                    "avatarPath": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    }
                },
                "type": "object",
                "example": {
                    "id": 1,
                    "firstName": "Test",
                    "lastName": "User",
                    "gender": "Female",
                    "phone": "+380501234567",
                    "currentAddress": "Beech Creek, PA, Pennsylvania",
                    "permanantAddress": "Arlington Heights, IL, Illinois",
                    "email": "12345@gmail.com",
                    "birthday": "1975-05-21",
                    "email_verified_at": "2023-11-06T15:34:03.000000Z",
                    "avatarPath": "avatar/127486g-2397y4-28397-2387h4.png",
                    "email_verified": false,
                    "created_at": "2023-11-06T15:34:03.000000Z",
                    "updated_at": "2023-11-06T15:34:03.000000Z"
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}